commit 1

+++ routes/web.php
@@ -17,27 +17,39 @@ Route::group(['namespace' => 'Main'], function() {
     Route::get('/', 'IndexController');
 });
 
-Route::group(['namespace' => 'Tag', 'prefix' => 'tags'], function() {
-    Route::get('/', 'IndexController')->name('tag.index');
-    Route::get('/create', 'CreateController')->name('tag.create');
-    Route::post('/', 'StoreController')->name('tag.store');
-    Route::get('/{tag}/edit', 'EditController')->name('tag.edit');
-    Route::patch('/{tag}', 'UpdateController')->name('tag.update');
-    Route::delete('/{tag}', 'DestroyController')->name('tag.destroy');
-});
-
 Route::group(['namespace' => 'Admin', 'prefix' => 'admin'], function () {
     Route::group(['namespace' => 'Main'], function() {
         Route::get('/', 'IndexController')->name('admin.main.index');
     });
+
+    Route::group(['namespace' => 'Post', 'prefix' => 'posts'], function() {
+        Route::get('/', 'IndexController')->name('admin.post.index');
+        Route::get('/create', 'CreateController')->name('admin.post.create');
+        Route::get('/{post}', 'ShowController')->name('admin.post.show');
+        Route::post('/', 'StoreController')->name('admin.post.store');
+        Route::get('/{post}/edit', 'EditController')->name('admin.post.edit');
+        Route::patch('/{post}', 'UpdateController')->name('admin.post.update');
+        Route::delete('/{post}', 'DestroyController')->name('admin.post.destroy');
+    });
+
     Route::group(['namespace' => 'Category', 'prefix' => 'categories'], function() {
         Route::get('/', 'IndexController')->name('admin.category.index');
         Route::get('/create', 'CreateController')->name('admin.category.create');
-        Route::get('/{category}}', 'ShowController')->name('admin.category.show');
+        Route::get('/{category}', 'ShowController')->name('admin.category.show');
         Route::post('/', 'StoreController')->name('admin.category.store');
         Route::get('/{category}/edit', 'EditController')->name('admin.category.edit');
         Route::patch('/{category}', 'UpdateController')->name('admin.category.update');
         Route::delete('/{category}', 'DestroyController')->name('admin.category.destroy');
+    });
+
+    Route::group(['namespace' => 'Tag', 'prefix' => 'tags'], function() {
+        Route::get('/', 'IndexController')->name('admin.tag.index');
+        Route::get('/create', 'CreateController')->name('admin.tag.create');
+        Route::get('/{tag}', 'ShowController')->name('admin.tag.show');
+        Route::post('/', 'StoreController')->name('admin.tag.store');
+        Route::get('/{tag}/edit', 'EditController')->name('admin.tag.edit');
+        Route::patch('/{tag}', 'UpdateController')->name('admin.tag.update');
+        Route::delete('/{tag}', 'DestroyController')->name('admin.tag.destroy');
     });
 });
